Fix the 3D surface visibility
Previously, the 3D surface checkbox would change the entire model's
visibility. This was misleading as the user would expect it to only control
the 3D view visibility.

Now the checkbox does control the visibility in the 3D view. To do so,
the display node is restrained to only the 2D view when we want to hide the
model annd unrestrained when we want to show it.